backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
pinecone-client==3.0.2
newsapi-python==0.2.7
requests==2.31.0
pydantic==2.6.1
python-multipart==0.0.9